Breadcrumb tests fixed for #86
