+++ packages/datagateway-common/src/state/actions/index.tsx
@@ -31,43 +31,43 @@ import axios from 'axios';
 import * as log from 'loglevel';
 import { fetchDownloadCart } from './cart';
 
-export const getApiFilter = (getState: () => StateType): URLSearchParams => {
-  const sort = getState().dgcommon.sort;
-  const filters = getState().dgcommon.filters;
+// export const getApiFilter = (getState: () => StateType): URLSearchParams => {
+//   const sort = getState().dgcommon.sort;
+//   const filters = getState().dgcommon.filters;
 
-  let searchParams = new URLSearchParams();
+//   let searchParams = new URLSearchParams();
 
-  for (let [key, value] of Object.entries(sort)) {
-    searchParams.append('order', JSON.stringify(`${key} ${value}`));
-  }
+//   for (let [key, value] of Object.entries(sort)) {
+//     searchParams.append('order', JSON.stringify(`${key} ${value}`));
+//   }
 
-  // sort by ID first to guarantee order
-  searchParams.append('order', JSON.stringify(`ID asc`));
+//   // sort by ID first to guarantee order
+//   searchParams.append('order', JSON.stringify(`ID asc`));
 
-  for (let [column, filter] of Object.entries(filters)) {
-    if (typeof filter === 'object') {
-      if ('startDate' in filter && filter.startDate) {
-        searchParams.append(
-          'where',
-          JSON.stringify({ [column]: { gte: `${filter.startDate} 00:00:00` } })
-        );
-      }
-      if ('endDate' in filter && filter.endDate) {
-        searchParams.append(
-          'where',
-          JSON.stringify({ [column]: { lte: `${filter.endDate} 23:59:59` } })
-        );
-      }
-    } else {
-      searchParams.append(
-        'where',
-        JSON.stringify({ [column]: { like: filter } })
-      );
-    }
-  }
+//   for (let [column, filter] of Object.entries(filters)) {
+//     if (typeof filter === 'object') {
+//       if ('startDate' in filter && filter.startDate) {
+//         searchParams.append(
+//           'where',
+//           JSON.stringify({ [column]: { gte: `${filter.startDate} 00:00:00` } })
+//         );
+//       }
+//       if ('endDate' in filter && filter.endDate) {
+//         searchParams.append(
+//           'where',
+//           JSON.stringify({ [column]: { lte: `${filter.endDate} 23:59:59` } })
+//         );
+//       }
+//     } else {
+//       searchParams.append(
+//         'where',
+//         JSON.stringify({ [column]: { like: filter } })
+//       );
+//     }
+//   }
 
-  return searchParams;
-};
+//   return searchParams;
+// };
 
 export * from './investigations';
 export * from './datasets';
@@ -76,27 +76,32 @@ export * from './cart';
 export * from './instruments';
 export * from './facilityCycles';
 
-export const sortTable = (
-  column: string,
-  order: Order | null
-): ActionType<SortTablePayload> => ({
-  type: SortTableType,
-  payload: {
-    column,
-    order,
-  },
-});
+// =============================================
+// export of sortTable and filterTable not working from index, moved to datafiles.tsx 
 
-export const filterTable = (
-  column: string,
-  filter: Filter | null
-): ActionType<FilterTablePayload> => ({
-  type: FilterTableType,
-  payload: {
-    column,
-    filter,
-  },
-});
+// export const sortTable = (
+//   column: string,
+//   order: Order | null
+// ): ActionType<SortTablePayload> => ({
+//   type: SortTableType,
+//   payload: {
+//     column,
+//     order,
+//   },
+// });
+
+// export const filterTable = (
+//   column: string,
+//   filter: Filter | null
+// ): ActionType<FilterTablePayload> => ({
+//   type: FilterTableType,
+//   payload: {
+//     column,
+//     filter,
+//   },
+// });
+
+// =================================
 
 // export const clearTable = (): Action => ({
 //   type: ClearTableType,
