+++ packages/datagateway-table/src/dls/tables/dlsDatafilesTable.component.tsx
@@ -16,9 +16,9 @@ import {
   fetchAllIds,
   sortTable,
   filterTable,
+  clearTable,
 } from 'datagateway-common';
 import { Paper, Typography } from '@material-ui/core';
-import { clearTable } from '../../state/actions';
 import { ThunkDispatch } from 'redux-thunk';
 import { connect } from 'react-redux';
 import { StateType } from '../../state/app.types';
