+++ packages/datagateway-table/src/dls/tables/dlsProposalsTable.component.test.tsx
@@ -10,12 +10,12 @@ import {
   sortTable,
   fetchInvestigationCountRequest,
   dGCommonInitialState,
+  clearTable,
 } from 'datagateway-common';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { MemoryRouter } from 'react-router';
 import axios from 'axios';
-import { clearTable } from '../../state/actions';
 
 describe('DLS Proposals table component', () => {
   let shallow;
