+++ packages/datagateway-table/src/isis/tables/isisDatafilesTable.component.tsx
@@ -19,10 +19,10 @@ import {
   fetchAllIds,
   sortTable,
   filterTable,
+  clearTable,
 } from 'datagateway-common';
 import { Paper, IconButton } from '@material-ui/core';
 import { GetApp } from '@material-ui/icons';
-import { clearTable } from '../../state/actions';
 import { ThunkDispatch } from 'redux-thunk';
 import { connect } from 'react-redux';
 import { StateType } from '../../state/app.types';
