+++ packages/datagateway-table/src/isis/tables/isisInvestigationsTable.component.test.tsx
@@ -13,13 +13,13 @@ import {
   removeFromCartRequest,
   addToCartRequest,
   dGCommonInitialState,
+  clearTable,
 } from 'datagateway-common';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { Table } from 'datagateway-common';
 import { MemoryRouter } from 'react-router';
 import axios from 'axios';
-import { clearTable } from '../../state/actions';
 
 describe('ISIS Investigations table component', () => {
   let shallow;
@@ -231,7 +231,7 @@ describe('ISIS Investigations table component', () => {
   });
 
   it('sends removeFromCart action on checked checkbox click', () => {
-    state.dgtable.cartItems = [
+    state.dgcommon.cartItems = [
       {
         entityId: 1,
         entityType: 'investigation',
@@ -259,7 +259,7 @@ describe('ISIS Investigations table component', () => {
   });
 
   it('selected rows only considers relevant cart items', () => {
-    state.dgtable.cartItems = [
+    state.dgcommon.cartItems = [
       {
         entityId: 2,
         entityType: 'investigation',
@@ -305,14 +305,14 @@ describe('ISIS Investigations table component', () => {
 
     const detailsPanelWrapper = shallow(
       wrapper.find(Table).prop('detailsPanel')({
-        rowData: state.dgtable.data[0],
+        rowData: state.dgcommon.data[0],
       })
     );
     expect(detailsPanelWrapper).toMatchSnapshot();
 
     mount(
       wrapper.find(Table).prop('detailsPanel')({
-        rowData: state.dgtable.data[0],
+        rowData: state.dgcommon.data[0],
         detailsPanelResize: jest.fn(),
       })
     );
@@ -361,8 +361,8 @@ describe('ISIS Investigations table component', () => {
   });
 
   it('gracefully handles missing Study from Study Investigation object and missing Instrument from InvestigationInstrument object', () => {
-    state.dgtable.data[0] = {
-      ...state.dgtable.data[0],
+    state.dgcommon.data[0] = {
+      ...state.dgcommon.data[0],
       INVESTIGATIONINSTRUMENT: [
         {
           ID: 1,
