+++ packages/datagateway-table/src/isis/tables/isisInvestigationsTable.component.tsx
@@ -18,13 +18,13 @@ import {
   fetchAllISISInvestigationIds,
   sortTable,
   filterTable,
+  clearTable,
 } from 'datagateway-common';
 import { StateType } from '../../state/app.types';
 import { connect } from 'react-redux';
 import { Action, AnyAction } from 'redux';
 import { TableCellProps, IndexRange } from 'react-virtualized';
 import { ThunkDispatch } from 'redux-thunk';
-import { clearTable } from '../../state/actions';
 import InvestigationDetailsPanel from '../detailsPanels/investigationDetailsPanel.component';
 import useAfterMountEffect from '../../utils';
