+++ packages/datagateway-table/src/table/datafileTable.component.test.tsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import { createShallow, createMount } from '@material-ui/core/test-utils';
 import DatafileTable from './datafileTable.component';
-import { initialState } from '../state/reducers/dgtable.reducer';
+import { initialState as dgTableInitialState } from '../state/reducers/dgtable.reducer';
 import configureStore from 'redux-mock-store';
-import { StateType, Datafile } from '../state/app.types';
-import {
-  fetchDatafilesRequest,
+import { StateType,  } from '../state/app.types';
+import {   fetchDatafilesRequest,
+  fetchAllIdsRequest,
+  clearTable,
+  Datafile,
   filterTable,
   sortTable,
   downloadDatafileRequest,
   addToCartRequest,
   removeFromCartRequest,
-  fetchDatafileCountRequest,
-  clearTable,
-  fetchAllIdsRequest,
-} from '../state/actions';
+  fetchDatafileCountRequest, } from 'datagateway-common';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { MemoryRouter } from 'react-router';
 import axios from 'axios';
+import {dGCommonInitialState,} from 'datagateway-common';
 
 describe('Datafile table component', () => {
   let shallow;
@@ -36,8 +36,8 @@ describe('Datafile table component', () => {
     mount = createMount();
 
     mockStore = configureStore([thunk]);
-    state = JSON.parse(JSON.stringify({ dgtable: initialState }));
-    state.dgtable.data = [
+    state = JSON.parse(JSON.stringify({ dgcommon: dGCommonInitialState, dgtable: dgTableInitialState }));
+    state.dgcommon.data = [
       {
         ID: 1,
         NAME: 'Test 1',
@@ -47,7 +47,7 @@ describe('Datafile table component', () => {
         DATASET_ID: 1,
       },
     ];
-    state.dgtable.allIds = [1];
+    state.dgcommon.allIds = [1];
   });
 
   afterEach(() => {
@@ -191,7 +191,7 @@ describe('Datafile table component', () => {
   });
 
   it('sends removeFromCart action on checked checkbox click', () => {
-    state.dgtable.cartItems = [
+    state.dgcommon.cartItems = [
       {
         entityId: 1,
         entityType: 'datafile',
@@ -219,7 +219,7 @@ describe('Datafile table component', () => {
   });
 
   it('selected rows only considers relevant cart items', () => {
-    state.dgtable.cartItems = [
+    state.dgcommon.cartItems = [
       {
         entityId: 1,
         entityType: 'dataset',
@@ -269,9 +269,9 @@ describe('Datafile table component', () => {
   });
 
   it("doesn't display download button for datafiles with no location", () => {
-    const datafile = state.dgtable.data[0] as Datafile;
+    const datafile = state.dgcommon.data[0] as Datafile;
     const { LOCATION, ...datafileWithoutLocation } = datafile;
-    state.dgtable.data = [datafileWithoutLocation];
+    state.dgcommon.data = [datafileWithoutLocation];
 
     const testStore = mockStore(state);
     const wrapper = mount(
@@ -293,7 +293,7 @@ describe('Datafile table component', () => {
     );
     const detailsPanelWrapper = shallow(
       wrapper.prop('detailsPanel')({
-        rowData: state.dgtable.data[0],
+        rowData: state.dgcommon.data[0],
       })
     );
     expect(detailsPanelWrapper).toMatchSnapshot();
