+++ packages/datagateway-table/src/table/datafileTable.component.tsx
@@ -19,9 +19,9 @@ import {
   fetchAllIds,
   sortTable,
   filterTable,
+  clearTable,
 } from 'datagateway-common';
 import { GetApp } from '@material-ui/icons';
-import { clearTable } from '../state/actions';
 import { ThunkDispatch } from 'redux-thunk';
 import { connect } from 'react-redux';
 import { StateType } from '../state/app.types';
