+++ packages/datagateway-table/src/table/datasetTable.component.tsx
@@ -17,8 +17,8 @@ import {
   fetchAllIds,
   sortTable,
   filterTable,
+  clearTable,
 } from 'datagateway-common';
-import { clearTable } from '../state/actions';
 import { AnyAction } from 'redux';
 import { StateType } from '../state/app.types';
 import { ThunkDispatch } from 'redux-thunk';
